Docs: Better JSONSerializer docs
